% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{offshore_datasets}
\alias{Addresses}
\alias{Entities}
\alias{Intermediaries}
\alias{Officers}
\alias{all_edges}
\alias{offshore_datasets}
\title{Datasets from the ICIJ Offshore Leaks Database}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 319150 rows and 21 columns.}
\source{
The International Consortium of Investigative Journalists. See here
for more: \url{https://offshoreleaks.icij.org/pages/database}
}
\usage{
Entities

Officers

Intermediaries

Addresses

all_edges
}
\description{
This contains five datasets provided as CSV files by the ICIJ Offshore Leaks
Database, as tbl_df objects (a form of data frame) that can easily be explored
in R.
}
\details{
This contains data about four types of nodes: Entities, Officers,
Intermediaries, and Addresses. For definitions of each of these, see
the ICIJ page here: \url{https://offshoreleaks.icij.org/pages/about#terms_definition}.
It also contains an \code{all_edges} dataset that links them together.
}
\examples{

library(dplyr)
Entities \%>\%
  count(countries, sort = TRUE)

Intermediaries \%>\%
  count(countries, sort = TRUE)

}
\keyword{datasets}

